ModelMapping applied

+++ src/main/java/com/kodlamaio/rentAcar/bussines/concretes/CarManager.java
@@ -1,6 +1,7 @@
 package com.kodlamaio.rentAcar.bussines.concretes;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -9,43 +10,33 @@ import com.kodlamaio.rentAcar.bussines.abstracts.CarService;
 import com.kodlamaio.rentAcar.bussines.request.cars.CreateCarRequest;
 import com.kodlamaio.rentAcar.bussines.request.cars.DeleteCarRequest;
 import com.kodlamaio.rentAcar.bussines.request.cars.UpdateCarRequest;
+import com.kodlamaio.rentAcar.bussines.response.cars.GetAllCarsResponse;
+import com.kodlamaio.rentAcar.core.utilities.mapping.ModelMapperService;
 import com.kodlamaio.rentAcar.core.utilities.result.DataResult;
 import com.kodlamaio.rentAcar.core.utilities.result.ErrorResult;
 import com.kodlamaio.rentAcar.core.utilities.result.Result;
 import com.kodlamaio.rentAcar.core.utilities.result.SuccessDataResult;
 import com.kodlamaio.rentAcar.core.utilities.result.SuccessResult;
 import com.kodlamaio.rentAcar.dataAccess.abstracts.CarRepository;
-import com.kodlamaio.rentAcar.entities.concretes.Brand;
 import com.kodlamaio.rentAcar.entities.concretes.Car;
-import com.kodlamaio.rentAcar.entities.concretes.Color;
 
 @Service
 public class CarManager implements CarService{
 	
 	private CarRepository carRepository;
+	private ModelMapperService modelMapperService;
 	
 	@Autowired
-	public CarManager(CarRepository carRepository) {
+	public CarManager(CarRepository carRepository,ModelMapperService modelMapperService) {
 		this.carRepository = carRepository;
+		this.modelMapperService = modelMapperService;
 	}
 	
 	@Override
 	public Result add(CreateCarRequest createCarRequest) {
 		if(ifCheckExist(createCarRequest.getBrandId())<5) {
-			Car car = new Car();
-			car.setDailyPrice(createCarRequest.getDailyPrice());
-			car.setDescription(createCarRequest.getDescription());
-			car.setPlate(createCarRequest.getPlate());
-			car.setKm(createCarRequest.getKm());
+			Car car = this.modelMapperService.forRequest().map(createCarRequest, Car.class);
 			car.setState(1);
-			
-			Brand brand = new Brand();
-			brand.setId(createCarRequest.getBrandId());
-			car.setBrand(brand);
-			
-			Color color = new Color();
-			color.setId(createCarRequest.getColorId());
-			car.setColor(color);
 			this.carRepository.save(car);
 			return new SuccessResult("CAR.ADDED");
 		}else {
@@ -56,35 +47,30 @@ public class CarManager implements CarService{
 
 	@Override
 	public Result delete(DeleteCarRequest deleteCarRequest) {
-		carRepository.deleteById(deleteCarRequest.getId());
-		return new SuccessResult("CAR.DELETE");
+		Car car = carRepository.findById(deleteCarRequest.getId());
+		if(car.getState()==1) {
+			carRepository.deleteById(deleteCarRequest.getId());
+			return new SuccessResult("CAR.DELETE");
+		}else {
+			return new ErrorResult("COULD.NOT.DELETE!!!");
+		}
+		
 	}
 
 	@Override
 	public Result update(UpdateCarRequest updateCarRequest) {
-		Car car = carRepository.findById(updateCarRequest.getId());
-		car.setDescription(updateCarRequest.getDescription());
-		car.setDailyPrice(updateCarRequest.getDailyPrice());
-		car.setPlate(updateCarRequest.getPlate());
-		car.setKm(updateCarRequest.getKm());
-		
-		Brand brand = new Brand();
-		brand.setId(updateCarRequest.getBrandId());
-		car.setBrand(brand);
-		
-		Color color = new Color();
-		color.setId(updateCarRequest.getColorId());
-		car.setColor(color);
-		
+		Car car = this.modelMapperService.forRequest().map(updateCarRequest, Car.class);
+		car.setState(1);
 		carRepository.save(car);
-		
 		return new SuccessResult("CAR.UPDATE");
 	}
 
 	@Override
-	public DataResult<List<Car>> getAll() {
-		
-		return new SuccessDataResult<List<Car>>(carRepository.findAll(),"CAR.LİSTED");
+	public DataResult<List<GetAllCarsResponse>> getAll() {
+		List<Car> cars = this.carRepository.findAll();
+		List<GetAllCarsResponse> response = cars.stream().map(car -> this.modelMapperService.forResponse()
+				.map(car, GetAllCarsResponse.class)).collect(Collectors.toList());
+		return new SuccessDataResult<List<GetAllCarsResponse>>(response);
 	}
 
 	@Override
